requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.11
packaging==26.0
pyserial==3.5
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
